Refactor: move all pages to /pages, add feed
