+++ src/App.tsx
@@ -1,14 +1,14 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { ThemeProvider } from '@mui/material';
 import theme from './components/theme';
 import { CssBaseline, Box } from '@mui/material';
 import Nav from './components/basic/Nav';
-import ChatPage from './ChatPage';
-import MatchesPage from './MatchesPage';
-import SearchPage from './SearchPage';
-import LoginPage from './LoginPage';
-import TestsPage from './TestsPage';
-import ProfilePage from './ProfilePage';
+import ChatPage from './components/pages/ChatPage';
+import MatchesPage from './components/pages/MatchesPage';
+import FeedPage from './components/pages/FeedPage';
+import LoginPage from './components/pages/LoginPage';
+import TestsPage from './components/pages/TestsPage';
+import ProfilePage from './components/pages/ProfilePage';
 import Messages from './Messages';
 import { Routes, Route, useLocation } from 'react-router-dom';
 
@@ -55,6 +55,27 @@ const contacts = [
   },
 ];
 
+const people = [
+  {
+    id: 1,
+    name: 'Jane Smith1',
+    description: 'Product Designer',
+    imageUrl: 'https://steamuserimages-a.akamaihd.net/ugc/1844789643806854188/FB581EAD503907F56A009F85371F6FB09A467FEC/?imw=512&imh=497&ima=fit&impolicy=Letterbox&imcolor=%23000000&letterbox=true',
+  },
+  {
+    id: 2,
+    name: 'Jane Smith2',
+    description: 'Product Designer',
+    imageUrl: 'https://i.pinimg.com/736x/56/21/7b/56217b1ef6a69a2583ff13655d48bc53.jpg',
+  },
+  {
+    id: 3,
+    name: 'Jane Smith3',
+    description: 'Product Designer',
+    imageUrl: 'https://avatars.yandex.net/get-music-content/5878680/7bee58da.a.25445174-1/m1000x1000?webp=false',
+  },
+];
+
 function App() {
   return (
     <ThemeProvider theme={theme}>
@@ -69,6 +90,12 @@ function AppContent() {
 
   const shouldHideNav = /^\/.+\/[^/]+$/.test(location.pathname);
 
+  const [currentIndex, setCurrentIndex] = useState(0);
+  const getNext = () => {
+    setCurrentIndex((prev) => (prev + 1) % people.length);
+    return people[currentIndex];
+  }
+
   return (
     <>
       <Box sx={{ pb: 7 }}>
@@ -76,7 +103,7 @@ function AppContent() {
           <Route path="/chats" element={<ChatPage contacts={contacts} />} />
           <Route path="/chat/:id" element={<Messages contacts={contacts} />} />
           <Route path="/matches" element={<MatchesPage />} />
-          <Route path="/search" element={<SearchPage />} />
+          <Route path="/search" element={<FeedPage getNextPerson={getNext} onLike={console.log} onDislike={console.log} onSuperLike={console.log} />} />
           <Route path="/tests" element={<TestsPage />} />
           <Route path="/profile" element={<ProfilePage />} />
           <Route path="*" element={<LoginPage />} />
